move stuff back into the ViewModel
